Add logic for full adder using 2 half adders.
However the implementation is not very intutive to me. Once we know that
sum can be made via a + b and then + c then we can refactor the sum
expression into thhat form... but I'm unsure how to get to that result
first - not sure if trail and error is the only method.

Also not sure why just or's the 2 carry just works.