Add implementation for ALU out output.
This can ofcourse be optimized a lot (maybe) One I can clearly see that
first layers of mux for x and y are not required and can be replaced
with and's with not of zx and zy respectively.

But there maybe more...

And its also hard to understand this HDL. I will like to see what others
have come up and how to make it more readble - if there is another way
than just extracting things in smaller units... But I have a feeling
that it'll still not be the same. But we can by extracting smaller chips
to say zx'd input. nx'd input.. blah blah.